+++ packages/web/src/app/layout.tsx
@@ -30,7 +30,7 @@ export default async function RootLayout({
   await connection();
 
   return (
-    <html lang="en">
+    <html lang="en" suppressHydrationWarning>
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >