+++ packages/web/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @workflow/web
 
+## 4.0.1-beta.4
+
+### Patch Changes
+
+- dbf2207: Web: refactor active/hover styles from trace viewer to avoid color conflicts
+- eadf588: Add button to re-run workflows
+
+## 4.0.1-beta.3
+
+### Patch Changes
+
+- 731adff: Fix run data not updating live on run detail view
+- 22917ab: Web: fix resource detail sidebar briefly showing old data when updating selection
+- 66225bf: Web: Allow filtering by workflow name and status on the runs list view
+- 9ba86ce: Web: fix links to docs
+
 ## 4.0.1-beta.2
 
 ### Patch Changes