+++ packages/world-postgres/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @workflow/world-postgres
 
+## 4.0.1-beta.3
+
+### Patch Changes
+
+- ae0972f: Fix build script to include the built files
+
+## 4.0.1-beta.2
+
+### Patch Changes
+
+- 7868434: Remove `AuthProvider` interface from `World` and associated implementations
+- Updated dependencies [d3a4ed3]
+- Updated dependencies [d3a4ed3]
+- Updated dependencies [66225bf]
+- Updated dependencies [7868434]
+  - @workflow/world@4.0.1-beta.2
+  - @workflow/world-local@4.0.1-beta.2
+
 ## 4.0.1-beta.1
 
 ### Patch Changes